requirements.txt
Django==1.4
certifi==0.0.8
chardet==1.0.1
httplib2==0.7.4
oauthlib==0.1.3
pyasn1==0.1.3
requests==0.12.1
rsa==3.0.1
wsgiref==0.1.2
gunicorn==0.14.2
-e git://github.com/brosner/python-oauth2.git#egg=oauth2
pybars==0.0.1
django-autoslug==1.5.0